+++ src/Event/EventType.php
@@ -7,4 +7,5 @@ enum EventType: string
     case ALL = "all";
     case PING = "ping";
     case ISSUE = "issue";
+    case PULL_REQUEST = "pull_request";
 }