+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0 - 2024-07-26
 
 - support for the event `issue`
+- support for the event `pull_request`
 
 ## 1.0.0 - 2024-07-25